+++ 20-cards/game_naive.h
@@ -10,16 +10,17 @@ using std::ostream;
 using std::vector;
 
 struct Card {
-  int c;
+  size_t c;
 };
+
 ostream &operator<<(ostream &os, const Card &c);
-inline int get_point(Card c);
-inline int get_suit(Card c);
+inline size_t get_point(Card c);
+inline size_t get_suit(Card c);
 
 using Game = array<Card, 52>;
 Game create_game();
 
 bool is_equal_point(Card c, Card d);
 bool is_equal_suit(Card c, Card d);
-bool is_serial(vector<Card> cards, int card_max = 13);
+bool is_serial(vector<Card> cards, size_t card_max = 13);
 } // namespace GameNaive
